requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
fastapi==0.115.5
filelock==3.16.1
h11==0.14.0
idna==3.10
packaging==24.2
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
snowflake-connector-python==3.12.3
snowflake-sqlalchemy==1.6.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
